Rename `ab_contracts_common::method` -> `ab_contracts_common::methods`
