+++ crates/contracts/ab-contracts-common/src/methods.rs
